suples/requirements.txt
Django==5.0.6
asgiref==3.8.1
django-cors-headers==4.4.0
djangorestframework==3.15.1
future==1.0.0
mysqlclient==2.2.4
pip==23.2.1
serializers==0.2.4
setuptools==68.2.0
sqlparse==0.5.0
typing_extensions==4.12.2
wheel==0.41.2
